creatin a different view outside of docker api space for crane repository viewing
